+++ app/(home)/layout.tsx
@@ -1,4 +1,4 @@
-import Navbar from '../components/Navbar';
+import Navbar from '../../components/Navbar';
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
