--- app/components/LoginButton.tsx
@@ -1,15 +1,0 @@
-'use client';
-
-import { signIn } from 'next-auth/react';
-
-export default function LoginButton() {
-  return (
-    <button
-      onClick={() => {
-        signIn();
-      }}
-    >
-      Login
-    </button>
-  );
-}
