--- app/components/LogoutButton.tsx
@@ -1,15 +1,0 @@
-'use client';
-
-import { signOut } from 'next-auth/react';
-
-export default function LogoutButton() {
-  return (
-    <button
-      onClick={() => {
-        signOut();
-      }}
-    >
-      Logout
-    </button>
-  );
-}
