+++ utils/getUserData.ts
@@ -3,15 +3,14 @@ import { getServerSession } from 'next-auth';
 
 const getUserData = async () => {
   const session = await getServerSession(authOptions);
+  if (!session || !session.user) {
+    return null;
+  }
 
-  const userData = {
-    name: session?.user?.name,
-    email: session?.user?.email,
+  const userData: UserData = {
+    name: session.user.name!,
+    email: session.user.email!,
   };
-  // console.log('userData: ', userData);
-  // console.log('==================================');
-  // console.log('session: ', session);
-  // console.log('==================================');
   return userData;
 };
